Moved startUploaderService() back to DataManager
The method will live here.

The tasks are about to be refactored. AbstractDataManagerTask will gain
a method that does the equivalent of all the EventBus....addEvents()
now, but will also run DataManager.startUploaderService().  I didn't
want to pollute EventBus with holding an application context and
overload it into starting services.